AM/src/androidTest/java/com/audiomack/ui/authentication/AuthenticationActivityTest.kt
package com.audiomack.ui.authentication

import androidx.test.espresso.Espresso.onView
import androidx.test.espresso.action.ViewActions.click
import androidx.test.espresso.assertion.ViewAssertions.matches
import androidx.test.espresso.intent.Intents
import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
import androidx.test.espresso.matcher.ViewMatchers.withId
import androidx.test.ext.junit.runners.AndroidJUnit4
import androidx.test.rule.ActivityTestRule
import com.audiomack.R
import org.junit.After
import org.junit.Assert.assertTrue
import org.junit.Before
import org.junit.Rule
import org.junit.Test
import org.junit.runner.RunWith

@RunWith(AndroidJUnit4::class)
class AuthenticationActivityTest {

    @get:Rule
    var activityTestRule = ActivityTestRule(AuthenticationActivity::class.java)

    @Before
    fun setup() {
        Intents.init()
    }

    @After
    fun clear() {
        Intents.release()
    }

    @Test
    fun visibility() {
        listOf(R.id.buttonClose).forEach {
            onView(withId(it)).check(matches(isDisplayed()))
        }
    }

    @Test
    fun close() {
        onView(withId(R.id.buttonClose)).perform(click())
        assertTrue(activityTestRule.activity.isFinishing)
    }
}
